Minor changes
added a new property in the csv file (malpensatrain).
replace a generic xcoord and ycoord properties name with a more
meaningful lat(itude) and lon(gitude) for the geographic coordinates of
each station.
Added a new png for the “malpensa train” service. You get this png if
you click on Garibaldi’s station.